requirements.txt
aiofiles==24.1.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
couchbase==4.3.5
dnspython==2.7.0
Flask==3.1.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
priority==2.0.0
python-dotenv==1.1.0
Quart==0.20.0
Werkzeug==3.1.3
wsproto==1.2.0
